requirements.txt
numpy
pandas
scikit-learn
scipy
torch>=2.2
pytorch-lightning==2.4.0  # optional
tensorboard
xgboost
prophet
omegaconf
hydra-core
matplotlib
seaborn
pyyaml
